\name{nonpar-package}
\alias{nonpar-package}
\alias{nonpar}
\docType{package}
\title{Collection of methods for non-parametric analysis
}
\description{Collection of methods for non-parametric analysis. At the moment only some tests for the two-sample location-scale problem are given, but more will be added in future versions.
}
\details{
\tabular{ll}{
Package: \tab nonpar\cr
Type: \tab Package\cr
Version: \tab 0.0-1\cr
Date: \tab 2015-02-26\cr
License: \tab GPL-2\cr
}

}
\author{
Theo Pepler

Maintainer: Theo Pepler <tpepler@sun.ac.za>
}
\references{
}
\keyword{ package }
\seealso{
}
\examples{
}
